-Added buffer effect interpolation for falloff
It is assumed that between audio buffers, effect parameters changed. This leads to a discontinuous jump between the amplitude at one buffer, and the amplitude at the next.
With linear interpolation, the audio buffer's samples are added such that they interpolate over the course of the buffer from the previous effect strength to the next.
Though this is currently just for experimental distance falloff, it will act as a prototype for additional audio buffer filters I may add.